requirements.txt
torch
torchvision
numpy
opencv-python
rasterio
Pillow
tqdm
matplotlib